ragnpg/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-huggingface>=0.0.6
psycopg2-binary>=2.9.9
pgvector>=0.2.5
sqlalchemy>=2.0.25
pypdf>=4.0.0
python-docx>=1.0.1
unstructured>=0.11.0
pdfplumber>=0.10.3
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
accelerate>=0.26.0
ollama>=0.1.6
python-dotenv>=1.0.0
tqdm>=4.66.1
pyyaml>=6.0.1 